chore: move initialisation strategies
